big split, merde_types has types, merde has macros.
