+++ merde/src/json_macros.rs
@@ -1,0 +1,31 @@
+#[doc(hidden)]
+#[macro_export]
+macro_rules! impl_json_serialize {
+    ($struct_name:ident < $lifetime:lifetime > { $($field:ident),+ }) => {
+        impl<$lifetime> $crate::merde_json::JsonSerialize for $struct_name<$lifetime> {
+            fn json_serialize(&self, serializer: &mut $crate::merde_json::JsonSerializer) {
+                #[allow(unused_imports)]
+                use $crate::MerdeError;
+
+                let mut guard = serializer.write_obj();
+                $(
+                    guard.pair(stringify!($field), &self.$field);
+                )+
+            }
+        }
+    };
+
+    ($struct_name:ident { $($field:ident),+ }) => {
+        impl $crate::merde_json::JsonSerialize for $struct_name {
+            fn json_serialize(&self, serializer: &mut $crate::merde_json::JsonSerializer) {
+                #[allow(unused_imports)]
+                use $crate::MerdeError;
+
+                let mut guard = serializer.write_obj();
+                $(
+                    guard.pair(stringify!($field), &self.$field);
+                )+
+            }
+        }
+    };
+}
