+++ merde_json/Cargo.toml
@@ -1,22 +1,18 @@
 [package]
 name = "merde_json"
-version = "2.4.1"
+version = "3.0.0"
 edition = "2021"
 authors = ["Amos Wenger <amos@bearcove.net>"]
-description = "Serialize and deserialize JSON with jiter and declarative macros"
+description = "JSON serialization and deserialization for merde, via jiter"
 license = "Apache-2.0 OR MIT"
 readme = "README.md"
-repository = "https://github.com/bearcove/merde_json"
+repository = "https://github.com/bearcove/merde"
 keywords = ["json", "serialization", "deserialization", "jiter"]
 categories = ["encoding", "parser-implementations"]
 
 [dependencies]
 jiter = "0.5.0"
-merde = { version = "0.1.0", path = "../merde" }
-smallvec = "1.13.2"
-
-[dev-dependencies]
-serde = { version = "1.0.204", features = ["derive"] }
+merde_types = { version = "3.0.0", path = "../merde_types" }
 
 [features]
 default = []
