--- merde_json_types/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-edition = "2021"
-name = "merde_json_types"
-version = "2.2.3"
-authors = ["Amos Wenger <amos@bearcove.net>"]
-description = "Wrapper types for merde_json that implement serialization/deserialization"
-license = "Apache-2.0 OR MIT"
-readme = "README.md"
-repository = "https://github.com/bearcove/merde_json"
-keywords = ["json", "serialization", "deserialization", "jiter", "merde"]
-categories = ["encoding", "parser-implementations"]
-
-[dependencies]
-merde_json = { version = "2.4.1", path = "../merde_json", optional = true }
-time = { version = "0.3.36", optional = true }
-
-[dev-dependencies]
-merde_json = { version = "2.4.1", path = "../merde_json" }
-time = { version = "0.3.36", features = ["macros"] }
-
-[features]
-default = []
-merde_json = ["dep:merde_json"]
-full = ["merde_json", "time-serialize", "time-deserialize"]
-time-types = ["dep:time"]
-time-serialize = ["time-types", "time/formatting"]
-time-deserialize = ["time-types", "time/parsing"]
