--- merde_json_types/src/lib.rs
@@ -1,5 +1,0 @@
-#![deny(missing_docs)]
-#![doc = include_str!("../README.md")]
-
-#[cfg(feature = "time-types")]
-pub mod time;
